fixed some literal bugs

+++ code/assignment_queue.cpp
@@ -287,7 +287,7 @@ int queueManger::reassign(FibonacciPQ *PQ)
  */
 int queueManger::doTreat(void)
 {
-    int total_num = assignment_list.size();
+    int total_num = (int) assignment_list.size();
     for (int i = 0; i < capacity; i++)
         if (locations[i])
             locations[i]->clear(&treated_list);
